pyproject.toml
[project]
name = "googlemaps-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.1.0",
    "nodriver>=0.48.1",
    "psycopg[binary]>=3.2.6",
    "pyyaml>=6.0.3",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
